wip - started to batch probability calculation

+++ source/algorithm/graph_processor.py
@@ -1,12 +1,16 @@
 import pickle
+from concurrent.futures import ProcessPoolExecutor
 from itertools import chain
 from pathlib import Path
+from typing import TypeVar, Generator
 
 from tqdm import tqdm
 
 from .graph_model import GraphModel
-from .utility import map_pool, print_stats
+from .utility import print_stats
 from .wrappers import GraphWrapper
+
+T = TypeVar('T')
 
 
 class GraphProcessor:
@@ -21,6 +25,9 @@ class GraphProcessor:
 
     # Processing pipeline
     __process_steps: list[callable]
+
+    # Step labels
+    __step_number: int
 
     def __init__(self,
                  epsilon: float,
@@ -41,39 +48,47 @@ class GraphProcessor:
         self.__process_steps = [
             GraphWrapper.load_file,
             GraphWrapper.preprocess,
-            self.__prune_graphs
+            self.prune_graph
         ]
 
-    def __prune_graphs(self, graph: GraphWrapper) -> GraphWrapper:
-        x = graph.prune(self.alpha, self.epsilon_p)
-        return x
+        self.__step_number = 0
+
+    def __step(self) -> str:
+        # Step counter for tqdm bars
+        self.__step_number += 1
+        return f'({self.__step_number})'
 
     def prune_from_paths(self, paths: list[Path]) -> list[tuple[GraphWrapper, int, int]]:
         previous = paths
         result: list = []
         for i, step in enumerate(self.__process_steps):
             step_label = f'({i+1}) {step.__name__.strip("_")}'
-            if self.args.single_threaded:
-                # Do a simple loop
-                result = [
-                    step(graph)
-                    for graph in tqdm(previous, step_label)
-                ]
-            else:
-                # Using map_pool pickles the arguments and results, which consumes a lot of memory
-                result = list(map_pool(
-                    step,
-                    previous,
-                    step_label
-                ))
+            result = list(self.map(
+                step,
+                previous,
+                step_label
+            ))
             del previous  # Release memory from de-pickled results
             previous = result
         return result
 
-    def perturb_graphs(self, paths: list[Path]) -> list[GraphWrapper]:
-        # Prune graphs
-        pruned_graphs_and_sizes_and_depths = self.prune_from_paths(paths)
-        # TODO: the fact that we're tracking stats IN the GraphWrapper
+    def process_graph(self, path: Path) -> tuple[list[GraphWrapper], list[tuple[str, GraphWrapper]]]:
+        # Apply n
+        result = path
+        for step in self.__process_steps:
+            result = step(result)
+        return result
+
+    def prune_graph(self, graph: GraphWrapper) -> GraphWrapper:
+        return graph.prune(self.alpha, self.epsilon_p)
+
+    def load_and_prune_graphs(self, paths: list[Path]) -> tuple[list[GraphWrapper], list[tuple[str, GraphWrapper]]]:
+        pruned_graphs_and_sizes_and_depths = list(self.map(
+            self.process_graph,
+            paths,
+            f'{self.__step()} Pruning graphs'
+        ))
+        # TODO: the fact that we're tracking stats in the GraphWrapper
         #       is a red flag that graph.prune should be in this class
         pruned_graphs = []
         sizes = []
@@ -91,21 +106,33 @@ class GraphProcessor:
         # Create training data and flatten into a single list
         train_data: list[tuple[str, GraphWrapper]] = list(
             chain.from_iterable(
-                map_pool(
+                self.map(
                     GraphWrapper.get_train_data,
                     pruned_graphs,
-                    '(4) Creating training data'
+                    f'{self.__step()} Creating training data'
                 )
             )
         )
+        return pruned_graphs, train_data
 
-        # Save a (pruned_graphs, training_data) tuple
-        with open('pruned_graphs_and_train_data.pkl', 'wb') as f:
-            pickle.dump((pruned_graphs, train_data), f)
-            print(f'  Wrote {len(pruned_graphs)} graphs and {len(train_data)} training samples to train_data.pkl')
+    def perturb_graphs(self, paths: list[Path]) -> list[GraphWrapper]:
+        pruned_graph_path = self.output_path / 'pruned_graphs.pkl'
+        if self.args.load_perturbed_graphs and pruned_graph_path.exists():
+            # Load graphs and training data from file
+            with open(pruned_graph_path, 'rb') as f:
+                pruned_graphs, train_data = pickle.load(f)
+                print(f'  Loaded {len(pruned_graphs)} graphs '
+                      f'and {len(train_data)} training samples from {pruned_graph_path}')
+        else:
+            # Perturb graphs and write graphs and training data to file
+            with open(pruned_graph_path, 'wb') as f:
+                pruned_graphs, train_data = self.load_and_prune_graphs(paths)
+                # Save a (pruned_graphs, training_data) tuple
+                pickle.dump((pruned_graphs, train_data), f)
+                print(f'  Wrote {len(pruned_graphs)} graphs and {len(train_data)} training samples to {pruned_graph_path}')
 
         # Train model
-        print('(5) Training model')
+        print(f'{self.__step()} Training model')
         paths = []
         graphs = []
         for path, graph in train_data:
@@ -118,23 +145,24 @@ class GraphProcessor:
              context_length=8,
              base_model_path=self.output_path / 'models'
         )
-        model.train(epochs=100)
+        model.train(epochs=1000)
 
         # Add graphs back (epsilon_m) # TODO: diff privacy here
         sizes = []
         num_marked_edges = []
         count_from_same_graph = []
-        for graph in tqdm(pruned_graphs, desc='(6) Re-attaching subgraphs'):
+        for graph in tqdm(pruned_graphs, desc=f'{self.__step()} Re-attaching subgraphs'):
             # Stats
             from_same_graph = 0
             num_marked_edges.append(len(graph.marked_edge_ids))
             # Re-attach a random to each marked edge
+            # for i in range(0, len(graph.marked_edge_ids), self.args.batch_size):
+
             for marked_edge_id, path in graph.marked_edge_ids.items():
                 # Predict a subgraph
                 subgraph = model.predict(path)
                 # Attach to the marked edge
                 graph.insert_subgraph(marked_edge_id, subgraph)
-
                 # Stats
                 sizes.append(len(subgraph))
                 if subgraph.source_edge_ref_id == graph.source_edge_ref_id:
@@ -149,3 +177,27 @@ class GraphProcessor:
         print()
         model.print_distance_stats()
         return pruned_graphs
+
+    def map(self,
+            func: callable,
+            items: list[T],
+            desc: str = '') -> Generator:
+        if self.args.single_threaded:
+            # Do a simple loop
+            for graph in tqdm(items, desc=desc):
+                yield func(graph)
+        else:
+            # Using this pickles the arguments and results, which consumes a lot of memory
+            with ProcessPoolExecutor() as executor:
+                # When we multiprocess, objects are pickled and copied in the child process
+                # instead of using the same object, so we have to return objects from the
+                # function to get the changes back
+                futures = [
+                    executor.submit(func, *item) if isinstance(item, tuple)
+                    else executor.submit(func, item)
+                    for item in items
+                ]
+                with tqdm(total=len(futures), desc=desc) as pbar:
+                    for future in futures:
+                        yield future.result()
+                        pbar.update(1)
